refactor: three components added
