+++ package.json
@@ -32,6 +32,7 @@
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^4.5.1",
     "identity-obj-proxy": "^3.0.0",
+    "jest": "^26.6.3",
     "node-sass": "^5.0.0",
     "react-app-polyfill": "^2.0.0",
     "react-refresh": "^0.9.0",
