+++ src/styles/content.scss
@@ -5,21 +5,6 @@
   margin: 0 auto;
 }
 
-header {
-  span.category {
-    display: block;
-    margin-top: 2rem;
-
-    font-size: 2.25rem;
-    font-weight: 600;
-    color: var(--gray);
-  }
-
-  span {
-    color: var(--white);
-  }
-}
-
 main {
   width: 100%;
   margin-top: 5.625rem;
