+++ src/styles/header.scss
@@ -1,0 +1,14 @@
+header {
+  span.category {
+    display: block;
+    margin-top: 2rem;
+
+    font-size: 2.25rem;
+    font-weight: 600;
+    color: var(--gray);
+  }
+
+  span {
+    color: var(--white);
+  }
+}
